[FIX] account: import of an init class was not working
a88027af5f5eb5d0efb4d2c42d5a179aa9bd2dc1 introduced a failed import
for the LegacyHTTPAdapter. Move the class into an own file and
import initialize.

Closes https://github.com/odoo/odoo/issues/126998
